Setup routing for public page
