+++ src/client/App/Dashboard/Navbar/index.js
