+++ src/client/App/Dashboard/index.js
@@ -1,0 +1,4 @@
+import Dashboard from "./Dashboard.jsx";
+import "./Dashboard.css";
+
+export default Dashboard;
