--- src/client/App/Home/Home.css
@@ -1,24 +1,0 @@
-.row {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    width: 100%;
-}
-
-.column {
-    display: flex;
-    flex-direction: column;
-    flex-basis: 100%;
-    flex: 1;
-    padding: 1rem 2rem;
-    overflow: scroll;
-}
-
-.column table {
-    border-top: 1px solid #ddd;
-    border-spacing: 0;
-}
-
-.column tbody tr:hover {
-    background-color: #f6f6f6;
-}
